% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpexploder.R
\name{bpexploder}
\alias{bpexploder}
\title{D3 Exploding Boxplots}
\usage{
bpexploder(data, settings = list(groupVar = NULL, levels = NULL, levelLabels =
  NULL, levelColors = NULL, yVar, yAxisLabel = NULL, xAxisLabel = NULL, tipText
  = NULL, referenceId = NULL, relativeWidth = 1), width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing all variables involved in the plot.}

\item{settings}{A list with the following elements:
\itemize{
 \item{\code{groupVar}}{ Character vector of length one setting name of grouping variable.
 (Leave NULL or omit for no grouping.)}
 \item{\code{levels}}{ Character vector of values of groupVar.  Leave NULL or omit to render
 levels in alphabetical order.}
 \item{\code{levelLabels}}{ Character vector setting custom labels for levels
 of grouping variable.}
 \item{\code{levelColors}}{ Character vector setting custom colors for boxes.}
 \item{\code{yVar}}{ Character vector of length one, specifying the numerical variable to plot.}
 \item{\code{yAxisLabel}}{ Character vector of length one.}
 \item{\code{xAxisLabel}}{ Character vector of length one.}
 \item{\code{tipText}}{ List of character vectors.  Names are variables to show in the
 tooltips, values are the labels to denote these variables.}
 \item{\code{referenceId}}{ Character vector; id attribute of an HTML tag.  Offset-width of
 this element is used to set width of the svg.  Defaults to the grandparent element.}
 \item{\code{relativeWidth}}{ Number between 0 and 1. The svg-width is this number
 multplied by the offset-width of the reference element.}
}}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.  Suggest to leave NULL.}

\item{elementId}{id of element ot contain widget.  Should be left NULL.}
}
\description{
Produce an svg element showing boxplots that explode upon mouse-click
into jittered individual-value plots.
}
